refactor(api): oRPC layer owns the registry and enrichment; add callOperation entry point
