+++ apps/erp/app/routes/api+/mcp+/lib/direct-executor.ts
@@ -1,138 +1,34 @@
-// Direct executor for ERP functions without MCP protocol wrapper
+// Direct executor for ERP functions without MCP protocol wrapper.
+//
+// TRANSITIONAL: the registry and the enrichment/extraction helpers now live in the
+// oRPC layer (`api+/v1+/lib/{registry,dispatch}.server.ts`) — this file re-imports
+// them so exactly one copy exists while its remaining callers migrate to
+// `callOperation`. It is deleted once the workflow dispatcher moves over.
 
 import type { Database } from "@carbon/database";
 import { getLogger } from "@carbon/logger";
 import type { SupabaseClient } from "@supabase/supabase-js";
-import * as accountFunctions from "~/modules/account/account.service";
-import * as accountingFunctions from "~/modules/accounting/accounting.ee.service";
-import * as documentsFunctions from "~/modules/documents/documents.service";
-import * as inventoryFunctions from "~/modules/inventory/inventory.service";
-import * as invoicingFunctions from "~/modules/invoicing/invoicing.service";
-import * as itemsFunctions from "~/modules/items/items.service";
-import * as peopleFunctions from "~/modules/people/people.service";
-import * as productionMcpFunctions from "~/modules/production/production.mcp.server";
-import * as productionFunctions from "~/modules/production/production.service";
-import * as purchasingFunctions from "~/modules/purchasing/purchasing.service";
-import * as qualityFunctions from "~/modules/quality/quality.service";
-import * as resourcesFunctions from "~/modules/resources/resources.service";
-import * as salesFunctions from "~/modules/sales/sales.service";
-import * as settingsFunctions from "~/modules/settings/settings.service";
-import * as sharedFunctions from "~/modules/shared/shared.service";
-import * as usersFunctions from "~/modules/users/users.service";
 import { getDatabaseClient } from "~/services/database.server";
+import {
+  enrichWithAuthContext,
+  extractOperation,
+  type McpOperation
+} from "../../v1+/lib/dispatch.server";
+import { functionRegistry } from "../../v1+/lib/registry.server";
 import { isMcpBlockedTool } from "./mcp-blocked-tools";
 import toolMetadata from "./tool-metadata.json";
 import type { AuthField } from "./types";
 
 const logger = getLogger("erp", "mcp", "direct-executor");
 
-// Combine all functions into a single registry. Exported so the oRPC dispatch
-// layer (`api+/v1+/lib/dispatch.server.ts`) resolves service functions the same way.
-export const functionRegistry = {
-  account: accountFunctions,
-  accounting: accountingFunctions,
-  documents: documentsFunctions,
-  inventory: inventoryFunctions,
-  invoicing: invoicingFunctions,
-  items: itemsFunctions,
-  people: peopleFunctions,
-  production: { ...productionFunctions, ...productionMcpFunctions },
-  purchasing: purchasingFunctions,
-  quality: qualityFunctions,
-  resources: resourcesFunctions,
-  sales: salesFunctions,
-  settings: settingsFunctions,
-  shared: sharedFunctions,
-  users: usersFunctions
-};
+export { enrichWithAuthContext, extractOperation, functionRegistry };
+export type { McpOperation };
 
 export interface ExecutorContext {
   client: SupabaseClient<Database>;
   companyId: string;
   companyGroupId: string;
   userId: string;
-}
-
-export type McpOperation = "create" | "update";
-
-// Stamps auth identity onto typed payloads. Carbon's services expect auth
-// fields inside the payload (predates MCP). `fields` is per-tool from
-// tool-metadata.json so reads stay clean and updates don't overwrite createdBy.
-export function enrichWithAuthContext(
-  value: unknown,
-  context: ExecutorContext,
-  fields: AuthField[],
-  operation?: McpOperation
-): unknown {
-  if (!value || typeof value !== "object") return value;
-  if (fields.length === 0) return value;
-
-  // Array payloads (e.g. the row list for upsertQuoteLinePrices) need per-element
-  // stamping — enrichment never reached inside them, so a NOT NULL createdBy on
-  // the row table failed. Only createdBy is injected into elements (and only for
-  // an insert): element keys are spread straight into an INSERT, so injecting
-  // companyId/updatedBy could add a column the row table doesn't have. The
-  // service owns companyId for these rows. createdBy is stamped AFTER the spread
-  // so a caller can't forge audit attribution by supplying it in a row.
-  if (Array.isArray(value)) {
-    if (operation === "update" || !fields.includes("createdBy")) return value;
-    return value.map((element) =>
-      element && typeof element === "object" && !Array.isArray(element)
-        ? { ...(element as Record<string, unknown>), createdBy: context.userId }
-        : element
-    );
-  }
-
-  const enriched: Record<string, unknown> = {
-    ...(value as Record<string, unknown>)
-  };
-
-  // A caller-supplied createdBy would send the service down its insert branch.
-  if (operation === "update") {
-    delete enriched.createdBy;
-  } else if (fields.includes("createdBy") && !("createdBy" in enriched)) {
-    enriched.createdBy = context.userId;
-  }
-  if (fields.includes("updatedBy")) {
-    enriched.updatedBy = context.userId;
-  }
-  if (fields.includes("companyId")) {
-    enriched.companyId = context.companyId;
-  }
-  if (fields.includes("companyGroupId")) {
-    enriched.companyGroupId = context.companyGroupId;
-  }
-
-  return enriched;
-}
-
-// Pulls the MCP-only `_operation` flag out of the args, top level or nested.
-// Returns every value it found so the caller can reject contradictory ones.
-export function extractOperation(args: Record<string, any> | undefined): {
-  operations: string[];
-  args: Record<string, any> | undefined;
-} {
-  if (!args) return { operations: [], args };
-
-  const operations: string[] = [];
-  const cleaned: Record<string, any> = {};
-
-  if (args._operation !== undefined) operations.push(String(args._operation));
-
-  for (const [key, value] of Object.entries(args)) {
-    if (key === "_operation") continue;
-    if (value && typeof value === "object" && !Array.isArray(value)) {
-      const { _operation, ...rest } = value as Record<string, any>;
-      if (_operation !== undefined) {
-        operations.push(String(_operation));
-        cleaned[key] = rest;
-        continue;
-      }
-    }
-    cleaned[key] = value;
-  }
-
-  return { operations, args: cleaned };
 }
 
 export async function executeFunction(
